Refactor User and Session route, controller, and service into an Auth route, controller, and service
